refactor: Playerの関心ごとをPlayerにまとめる
